LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/textfieldselection/selection-start-end-expected.txt


Harness Error (TIMEOUT), message = null

PASS Sanity check for testValue length; if this fails, variou absolute offsets in the test below need to be adjusted to be less than testValue.length 
FAIL onselect should fire when selectionStart is changed assert_equals: Initial .value set on textarea-appended should set selectionStart to end of value expected 10 but got 0
NOTRUN onselect should fire when selectionStart is changed on input-appended 
NOTRUN onselect should fire when selectionStart is changed on input-not-appended 
NOTRUN onselect should fire when selectionStart is changed on input-appended-prefocused 
NOTRUN onselect should fire when selectionStart is changed on input-not-appended-prefocused 
FAIL onselect should fire when selectionEnd is changed assert_equals: Initial .value set on textarea-appended should set selectionEnd to end of value expected 10 but got 0
NOTRUN onselect should fire when selectionEnd is changed on input-appended 
NOTRUN onselect should fire when selectionEnd is changed on input-not-appended 
NOTRUN onselect should fire when selectionEnd is changed on input-appended-prefocused 
NOTRUN onselect should fire when selectionEnd is changed on input-not-appended-prefocused 
FAIL Setting selectionStart to a value larger than selectionEnd should increase selectionEnd assert_equals: Initial .value set on textarea-appended should set selectionStart to end of value expected 10 but got 0
FAIL Setting selectionEnd to a value smaller than selectionStart should decrease selectionStart assert_equals: Initial .value set on textarea-appended should set selectionStart to end of value expected 10 but got 0
FAIL selectionStart edge-case values assert_equals: selectionStart setter on input-appended should convert -1 to 2^32-1 expected 10 but got 0
FAIL selectionEnd edge-case values assert_equals: selectionEnd setter on input-appended should convert -1 to 2^32-1 expected 10 but got 0
